fix depends functions
